Pull out all test in suits
