+++ jest.config.ts
@@ -1,6 +1,6 @@
 module.exports = {
     testEnvironment: "jsdom",
-    testResultsProcessor: "./postProcessor.js",
+    testResultsProcessor: "./postProcessor.mjs",
     moduleNameMapper: {
         "next/font/google": "<rootDir>/__mocks__/next/font/google.js",
         "^@/.*\\.(css|less|scss|sass)$": "<rootDir>/__mocks__/fileMock.js",
