+++ timing.json
@@ -1,24 +1,29 @@
 {
-    "testResults": [
-        {
-            "name": "src/__tests__/helper.test.ts",
-            "startTime": 1756472016566,
-            "endTime": 1756472016626
-        },
-        {
-            "name": "src/__tests__/temp-3.test.ts",
-            "startTime": 1756717443502,
-            "endTime": 1756717443544
-        },
-        {
-            "name": "src/__tests__/temp-1.test.ts",
-            "startTime": 1756717443549,
-            "endTime": 1756717443559
-        },
-        {
-            "name": "src/__tests__/temp-2.test.ts",
-            "startTime": 1756717443564,
-            "endTime": 1756717443583
-        }
-    ]
+  "testResults": [
+    {
+      "name": "src/__tests__/currentFormattedDate.test.ts",
+      "startTime": 1756728172199,
+      "endTime": 1756728172249
+    },
+    {
+      "name": "src/__tests__/packDataWithSequenceNumber.test.ts",
+      "startTime": 1756728172255,
+      "endTime": 1756728172271
+    },
+    {
+      "name": "src/__tests__/extractSequenceNumber.test.ts",
+      "startTime": 1756728172277,
+      "endTime": 1756728172292
+    },
+    {
+      "name": "src/__tests__/convertBytes.test.ts",
+      "startTime": 1756728172295,
+      "endTime": 1756728172308
+    },
+    {
+      "name": "src/__tests__/findMissingSequenceNumber.test.ts",
+      "startTime": 1756728172311,
+      "endTime": 1756728172326
+    }
+  ]
 }
