+++ tsconfig.json
@@ -19,7 +19,7 @@
             }
         ],
         "paths": {
-            "@/*": ["./src/*", "./scripts/*"]
+            "@/*": ["./src/*", "./scripts/*", "./postProcessor.mjs"]
         }
     },
     "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
